🐛 better counter
given dynamic interaction of the project, with sleep() prevention for button presses, it was possible with the previous implementation for a certain time-instance to get skipped completely.

for example if time's 299, and there's condition for (time % 300 == 0), but before the next increment, a button is pressed which brings in a few seconds of sleep with it, resulting in next time > 300.